requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Data Processing
pandas==2.1.1
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Database
psycopg2-binary==2.9.7

# Web Scraping
selenium==4.15.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and Configuration
python-dotenv==1.0.0

# Optional: For enhanced dashboard features
plotly==5.17.0
dash==2.14.1